Rewrite "info variables" to support line feeds and spaces in values, array keys and array values. (#23)
Bash's "typeset -p" escapes line feeds and other special characters and wraps such values in $''. Array keys with spaces are wrapped into "".
Previously, $'' was kept but the escape codes were resolved. The escape codes are now retained in the output of "info variables" to show each variable and its value on a single line. This is to improve readability and help programmatic parsing of the variable listing.